+++ fuc/api/common.py
@@ -95,6 +95,9 @@ def load_dataset(name, force=False):
             'tcga_laml.vcf',
             'tcga_laml_vep.vcf',
         ],
+        'brca': [
+            'brca.maf.gz',
+        ],
         'pyvcf': [
             'plot_comparison.vcf',
             'normal-tumor.vcf',